+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Unreleased
 
+-   Implement access to the outer structure.
+
+    See [`examples/02.rs`](./examples/02.rs) for the full example.
+
+    ```rust
+    rcstruct::rcstruct! {
+        pub struct GUI {}
+
+        impl {
+            pub new() -> Rt<Self> {
+                {}
+            }
+
+            pub fn window<T>(&self, data: T) -> WindowBuilder<T> {
+                let gui = outer().unwrap(); // <= `outer()` returns Option<GUI>
+                WindowBuilder { gui, data }
+            }
+        }
+    }
+    ```
+
 # 0.1.2 - 2019-06-08
 
 -   Identical to 0.1.1, but with correct changelog. :)
@@ -7,8 +28,6 @@
 # 0.1.1 - 2019-06-08
 
 -   Implement generic methods.
-
-    See [`examples/02.rs`](./examples/02.rs) for the full example.
 
     ```rust
     rcstruct::rcstruct! {